Refactor for artifacts demo
