+++ src/components/chat-with-artifacts/index.jsx
@@ -7,7 +7,7 @@ import { Flex } from '@wordpress/components';
 /**
  * Internal dependencies
  */
-import ChatHistory from './chat-history.jsx';
+import ChatHistory from '../chat-history.jsx';
 import ArtifactList from './artifact-list.jsx';
 import { useSelect } from '@wordpress/data';
 import useChatSettings from '../../hooks/use-chat-settings.js';
