+++ src/components/single-assistant-demo-ui.jsx
@@ -31,9 +31,9 @@ const SingleAssistantDemoUI = () => {
 
 	return (
 		<>
+			<ChatHistory />
 			<AgentUI />
 			<PopUpControls />
-			<ChatHistory />
 		</>
 	);
 };
